Moves nexus to interfaces
